+++ changelog.md
@@ -1,1 +1,1 @@
-- Updated to 1.18
+- Fixed blocks not saving